requirements.txt
Jinja2>=2.9.5
MarkupSafe==0.23
Werkzeug>=0.14
click==6.7
itsdangerous==0.24
requests==2.20.0
plaid-python
python-dotenv